requirements-api.txt
telethon
PyQt5
asyncio
emoji
python-dotenv
flask>=2.0.0
flask-cors>=3.0.10
flask-socketio>=5.1.1
requests>=2.25.1
python-dotenv>=0.19.0